v 1.0.0/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-jazzmin
django-cors-headers
gunicorn

requests

python-dotenv